requirements.txt
pdf2image
pillow
pytesseract
